Update gradle and plugin versions, add new changelog entry
Upgraded the version of gradle, plugins and JVM target for Kotlin. This update includes the latest features and improvements, providing better performance and stability.

A new entry in the changelog has been added to acknowledge the contribution of "rolandcheck" and his new feature: a contextual action to automatically create a mediatR handler.

Lastly, we added instructions to run the project on macOS to the README.md as there were previously no explicit instructions for macOS users.