Update JDK source option in build stage
The JDK source option in the build.yaml file was updated from 'AzureStorage' to 'PreInstalled'. This was done as the previous configuration was causing build instability issues. Instead of pulling the JDK from Azure storage during every build, the JDK is now pre-installed on the build servers. This will hopefully reduce build times and potential for errors. References to the Azure storage have been commented out for future reference if needed.