"Switch JDK source from PreInstalled to AzureStorage"
In the build configuration file, the JDK sourceOption parameter has been changed. Previously, the build script used a pre-installed version of JDK. Now, it fetches JDK from AzureStorage. This change ensures the build process always uses the identical, up-to-date version of JDK, which is stored in a specified AzureStorage account. This strategy improves the reproducibility and reliability of our builds.