Update jdkDirectory in build.yaml
Modified the jdkDirectory path in build.yaml file of CI/CD stages in the DevOps. The change removes the '/bin' segment from the end of the path. This modification is necessary to ensure that the system correctly identifies the Java Development Kit's (JDK) location. Deployments were failing previously due to this error in the path.